Reverted the KorinLoop's timeStep to be a float to not require all Component's to support doubles or suffer the performance cost of conversions.

+++ packages/korin/include/korin/system.h
@@ -18,7 +18,7 @@ class System
 {
 public:
    /// Sends the time step to update the Component and potentially its siblings.
-   virtual void update(double timeStep, const ComponentPtr& component) = 0;
+   virtual void update(float timeStep, const ComponentPtr& component) = 0;
 
    /// Notifies the System that a Component needs to be updated.
    virtual void notify(const ComponentPtr& component) = 0;
